backend/app/requirements.txt
# Framework
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Firebase
firebase-admin>=6.2.0

# HTTP client
aiohttp>=3.8.6
backoff>=2.2.1  # For rate limiting and retry logic

# LLaMA integration
tenacity>=8.2.3  # For retry mechanisms
ratelimit>=2.2.1  # For API rate limiting

# Image processing
Pillow>=10.0.1
numpy>=1.25.2
opencv-python>=4.8.1.78
tensorflow>=2.14.0
# For TensorFlow Lite (optional)
# tflite-runtime>=2.14.0

# Utilities
loguru>=0.7.2